+++ persistence/src/main/java/group10/persistence/model/Submission.java
@@ -10,12 +10,12 @@ public final class Submission {
     private final UUID roundId;
     private final String submissionPayload;
     private final Instant submittedAt;
-    private final Integer timeMs; // nullable
+    private final Long timeMs; // nullable
     private final boolean isCorrect;
     private final short score;
 
     public Submission(UUID id, UUID playerId, UUID matchId, UUID roundId, String submissionPayload,
-                      Instant submittedAt, Integer timeMs, boolean isCorrect, short score) {
+                      Instant submittedAt, Long timeMs, boolean isCorrect, short score) {
         this.id = id;
         this.playerId = playerId;
         this.matchId = matchId;
@@ -33,7 +33,7 @@ public final class Submission {
     public UUID getRoundId() { return roundId; }
     public String getSubmissionPayload() { return submissionPayload; }
     public Instant getSubmittedAt() { return submittedAt; }
-    public Integer getTimeMs() { return timeMs; }
+    public Long getTimeMs() { return timeMs; }
     public boolean isCorrect() { return isCorrect; }
     public short getScore() { return score; }
 }
